Finish adding authorizations to controller methods

+++ app/Http/Controllers/OrganizationController.php
@@ -30,6 +30,8 @@ class OrganizationController extends Controller
      */
     public function index()
     {
+        // Verify the user can access this function via policy
+        $this->authorize('viewAll', Organization::class);
 
         return response()->json(
             [
@@ -47,6 +49,9 @@ class OrganizationController extends Controller
      */
     public function store(OrganizationRequest $request)
     {
+        // Verify the user can access this function via policy
+        $this->authorize('create', Organization::class);
+
         if ($request->filled('id')) {
             $organization = Organization::find($request->id);
 
@@ -114,6 +119,9 @@ class OrganizationController extends Controller
         OrganizationRequest $request,
         Organization $organization
     ) {
+        // Verify the user can access this function via policy
+        $this->authorize('update', $organization);
+
         $owner = $organization->owner();
         $owner->update([
             'username'      => $request->username,
@@ -168,6 +176,9 @@ class OrganizationController extends Controller
      */
     public function destroy(Organization $organization)
     {
+        // Verify the user can access this function via policy
+        $this->authorize('delete', $organization);
+
         $owner = $organization->owner();
         $owner->delete();
         $organization->delete();
